added separator constant; added file management for each advert by user; fixed login;

+++ anunturi/controllers/advert.php
@@ -32,7 +32,7 @@ class Advert extends MY_CONTROLLER
 				'type' => $this->input->post('type'),
 				'date' => date('Y-m-d H:i:s'),
 		);
-		
+
 		$advertFile['advert_id'] = $this->advert_model->insert($advertData);
 
 		$files = scandir(UPLOAD_DIR);
@@ -40,8 +40,9 @@ class Advert extends MY_CONTROLLER
 		foreach ($files as $file)
 		{
 			if(in_array($file, array('.','..'))) continue;
-			$data['filename'] = $dest.$file;
-			copy(UPLOAD_DIR.$file, $data['filename']);
+			if(!$this->isFileForThisAdvert($file)) continue;
+			$data['filename'] = $file;
+			copy(UPLOAD_DIR.$file, $dest.$data['filename']);
 			unlink(UPLOAD_DIR.$file);
 			$advertFile['file_id'] = $this->files_model->insert($data);
 			$this->advert_files_model->insert($advertFile);
@@ -69,5 +70,13 @@ class Advert extends MY_CONTROLLER
 		$this->data['advert_files'] = $this->advert_files_model->getFilesForAdvert($advertId);
 		$this->data['user_data'] = $this->user_model->getById($this->data['advert']->user_id);
 		$this->loadView('advert_presentation', $this->data);
+	}
+
+	private function isFileForThisAdvert($file)
+	{
+		$advertGuid = $this->session->userdata('advert_guid');
+		$userId = $this->session->userdata('user_id');
+		$cmp = explode(SEP, $file);
+		return $cmp[1] == $userId && $cmp[2] == $advertGuid;
 	}
 }
